feat: implement rate strategy assemble

+++ big-market-infrastructure/src/main/java/cn/bugstack/infrastructure/persistent/po/StrategyRule.java
@@ -7,7 +7,7 @@ import java.util.Date;
 @Data
 public class StrategyRule {
     private Long id;
-    private Integer strategyId;
+    private Long strategyId;
     private Integer awardId;
     private Short ruleType;
     private String ruleModel;
